refactor: Mapper 함수를 Presentation Layer로 옮기기
