--- app/src/main/java/nextstep/shoppingcart/data/mapper/ProductMapper.kt
@@ -1,22 +1,0 @@
-package nextstep.shoppingcart.data.mapper
-
-import nextstep.shoppingcart.data.model.ProductEntity
-import nextstep.shoppingcart.model.Product
-
-fun Product.toEntity(): ProductEntity {
-    return ProductEntity(
-        id = id,
-        imageUrl = imageUrl,
-        name = name,
-        price = price
-    )
-}
-
-fun ProductEntity.toUi(): Product {
-    return Product(
-        id = id,
-        imageUrl = imageUrl,
-        name = name,
-        price = price,
-    )
-}
