+++ app/src/main/java/nextstep/shoppingcart/data/repository/CartRepository.kt
@@ -3,11 +3,9 @@ package nextstep.shoppingcart.data.repository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import nextstep.shoppingcart.data.datasource.CartLocalDataSource
-import nextstep.shoppingcart.data.mapper.toEntity
-import nextstep.shoppingcart.data.mapper.toUi
+import nextstep.shoppingcart.data.model.CartItemEntity
+import nextstep.shoppingcart.data.model.ProductEntity
 import nextstep.shoppingcart.data.util.SingletonHolder
-import nextstep.shoppingcart.model.CartItem
-import nextstep.shoppingcart.model.Product
 
 /**
  * 구현을 간소화하기 위해 interface 대신 구현체 바로 사용
@@ -16,22 +14,20 @@ class CartRepository private constructor(
     private val cartLocalDataSource: CartLocalDataSource,
 ) {
 
-    fun getItems(): Flow<List<CartItem>> = cartLocalDataSource.itemsFlow.map { items ->
-        items.map {
-            it.toUi()
-        }
+    fun getItems(): Flow<List<CartItemEntity>> = cartLocalDataSource.itemsFlow.map { items ->
+        items
     }
 
-    fun addOne(product: Product) {
-        cartLocalDataSource.addOne(product.toEntity())
+    fun addOne(product: ProductEntity) {
+        cartLocalDataSource.addOne(product)
     }
 
-    fun removeOne(product: Product) {
-        cartLocalDataSource.removeOne(product.toEntity())
+    fun removeOne(product: ProductEntity) {
+        cartLocalDataSource.removeOne(product)
     }
 
-    fun removeAll(product: Product) {
-        cartLocalDataSource.removeAll(product.toEntity())
+    fun removeAll(product: ProductEntity) {
+        cartLocalDataSource.removeAll(product)
     }
 
     companion object : SingletonHolder<CartRepository>(
