+++ app/src/main/java/nextstep/shoppingcart/data/repository/ProductRepository.kt
@@ -1,17 +1,14 @@
 package nextstep.shoppingcart.data.repository
 
 import nextstep.shoppingcart.data.datasource.ProductRemoteDataSource
-import nextstep.shoppingcart.data.mapper.toUi
+import nextstep.shoppingcart.data.model.ProductEntity
 import nextstep.shoppingcart.data.util.SingletonHolder
-import nextstep.shoppingcart.model.Product
 
 class ProductRepository private constructor(
     private val productRemoteDataSource: ProductRemoteDataSource,
 ) {
-    suspend fun fetch(): List<Product> {
-        return productRemoteDataSource.fetch().map {
-            it.toUi()
-        }
+    suspend fun fetch(): List<ProductEntity> {
+        return productRemoteDataSource.fetch()
     }
 
     companion object : SingletonHolder<ProductRepository>(
