+++ app/src/main/java/nextstep/shoppingcart/ui/product_detail/ProductDetailState.kt
@@ -1,6 +1,6 @@
 package nextstep.shoppingcart.ui.product_detail
 
-import nextstep.shoppingcart.model.Product
+import nextstep.shoppingcart.ui.model.Product
 
 data class ProductDetailState(
     val product: Product = Product("", "", "", 0),
